src/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
prometheus-client==0.19.0
pika==1.3.2
